feat: 컴포넌트 분리 및 mokdata 추가
