+++ components/index.ts
@@ -13,3 +13,5 @@ export { default as ModalTitle } from './modal/ModalTitle';
 export { default as Portal } from './modal/Portal';
 export { default as Modal } from './modal/Modal';
 export { default as CreateDashboardModal } from './modal/CreateDashboardModal';
+export { default as BoardList } from './myBoard/DashBoards';
+export { default as InvitationsDashBoards } from './myBoard/InvitationsDashBoards';
