+++ types/dashboards.ts
@@ -11,5 +11,5 @@ export interface Dashboards {
 export interface DashboardsRawData {
   cursorId: number;
   totalCount: number;
-  dashboards: Dashboards;
+  dashboards: Dashboards[];
 }
